+++ src/main/java/de/mossgrabers/convertwithmoss/format/nki/type/kontakt5/FileList.java
@@ -20,14 +20,9 @@ import java.util.List;
  */
 public class FileList
 {
-    private final List<String> sampleFilePaths = new ArrayList<> ();
-    private String             absoluteSourcePath;
-    private String             absoluteMonolithSourcePath;
-    private String             resourceFile;
-    private String             nkiFilename;
-    private String             wallpaperFilename;
-    private String             impulseResponseSampleFilename;
-    private String             resourceFolder;
+    private List<String> specialFiles;
+    private List<String> sampleFiles;
+    private List<String> otherFiles;
 
 
     /**
@@ -51,7 +46,7 @@ public class FileList
             if (version != 2)
                 throw new IOException (Functions.getMessage ("IDS_NKI5_UNSUPPORTED_FILELIST_EX_VERSION", Integer.toString (version)));
 
-            this.readSpecialFiles (in);
+            this.specialFiles = readFiles (in);
         }
         else
         {
@@ -60,52 +55,21 @@ public class FileList
                 throw new IOException (Functions.getMessage ("IDS_NKI5_UNSUPPORTED_FILELIST_VERSION", Long.toString (version)));
 
             if (version == 1)
-                this.absoluteMonolithSourcePath = readFile (in);
-        }
-
-        this.readFiles (in);
-        this.readMetadata (in, chunkID);
-    }
-
-
-    private void readSpecialFiles (final ByteArrayInputStream in) throws IOException
-    {
-        while (in.available () > 0)
-        {
-            final int type = StreamUtils.readSigned32 (in, false);
-            switch (type)
             {
-                case 0:
-                    this.absoluteSourcePath = readFile (in);
-                    return;
-
-                // Original absolute folder path only set for monolith
-                case 1:
-                    this.absoluteMonolithSourcePath = readFile (in);
-                    return;
-
-                // NKR resource file
-                case 2:
-                    this.resourceFile = readFile (in);
-                    this.resourceFolder = readFile (in);
-                    return;
-
-                // NKR resource file
-                case 3:
-                    this.resourceFile = readFile (in);
-                    break;
-
-                default:
-                    throw new IOException (Functions.getMessage ("IDS_NKI5_UNKNOWN_FILE_TYPE", Integer.toString (type)));
+                // absoluteMonolithSourcePath not used currently
+                readFile (in);
             }
         }
+
+        this.sampleFiles = readFiles (in);
+        this.readMetadata (in, chunkID);
     }
 
 
     private void readMetadata (final ByteArrayInputStream in, final int chunkID) throws IOException
     {
         // Last changed date of files
-        final int numFiles = this.sampleFilePaths.size ();
+        final int numFiles = this.sampleFiles.size ();
         for (int i = 0; i < numFiles; i++)
         {
             // The last update date of the file
@@ -120,37 +84,7 @@ public class FileList
             for (int i = 0; i < numFiles; i++)
                 StreamUtils.readUnsigned32 (in, false);
 
-            // Reverb samples (and maybe wallpaper)
-            while (in.available () > 0)
-            {
-                final int type = StreamUtils.readSigned32 (in, false);
-                switch (type)
-                {
-                    // There are no additional files at all
-                    case 0:
-                        return;
-
-                    // No more files
-                    case 1:
-                        return;
-
-                    case 2:
-                        this.nkiFilename = readFile (in);
-                        break;
-
-                    case 3:
-                        this.nkiFilename = readFile (in);
-                        this.wallpaperFilename = readFile (in);
-                        break;
-
-                    case 7:
-                        this.impulseResponseSampleFilename = readFileShort (in);
-                        break;
-
-                    default:
-                        throw new IOException (Functions.getMessage ("IDS_NKI5_UNKNOWN_FILE_TYPE", Integer.toString (type)));
-                }
-            }
+            this.otherFiles = readFiles (in);
         }
         else
         {
@@ -161,16 +95,22 @@ public class FileList
 
 
     /**
-     * Read all file names with their paths.
+     * Read several file paths.
      *
-     * @param in The data to parse
-     * @throws IOException Could not read the data
+     * @param in The input stream to read from
+     * @return The read file paths
+     * @throws IOException Could not read
      */
-    private void readFiles (final ByteArrayInputStream in) throws IOException
+    private static List<String> readFiles (final ByteArrayInputStream in) throws IOException
     {
-        final long fileCount = StreamUtils.readUnsigned32 (in, false);
-        for (long i = 0; i < fileCount; i++)
-            this.sampleFilePaths.add (readFile (in));
+        final List<String> files = new ArrayList<> ();
+        if (in.available () > 0)
+        {
+            final int size = StreamUtils.readSigned32 (in, false);
+            for (int i = 0; i < size; i++)
+                files.add (readFile (in));
+        }
+        return files;
     }
 
 
@@ -192,30 +132,13 @@ public class FileList
 
 
     /**
-     * Read one file name with it's path. The segment count is only 1 byte.
-     *
-     * @param in The data to parse
-     * @return The read file
-     * @throws IOException Could not read the data
-     */
-    private static String readFileShort (final ByteArrayInputStream in) throws IOException
-    {
-        final StringBuilder sb = new StringBuilder ();
-        final int segmentCount = in.read ();
-        for (int s = 0; s < segmentCount; s++)
-            readSegment (in, sb);
-        return sb.toString ();
-    }
-
-
-    /**
      * Reads a segment of a file path.
      *
      * @param in Where to read the segment from
-     * @param sb Where to append the read path segment
+     * @param buffer Where to append the read path segment
      * @throws IOException Could not read
      */
-    private static void readSegment (final ByteArrayInputStream in, final StringBuilder sb) throws IOException
+    private static void readSegment (final ByteArrayInputStream in, final StringBuilder buffer) throws IOException
     {
         final int segmentType = in.read ();
         switch (segmentType)
@@ -224,40 +147,63 @@ public class FileList
             case 0:
                 final String driveOld = StreamUtils.readASCII (in, 2).trim ();
                 if (driveOld.isEmpty ())
-                    sb.append ("/");
+                    buffer.append ("/");
                 else
-                    sb.append (driveOld).append (":/");
+                    buffer.append (driveOld).append (":/");
                 break;
 
             // Drive letter UTF-16
             case 1:
                 final String drive = StreamUtils.readWithLengthUTF16 (in);
                 if (drive.isEmpty ())
-                    sb.append ("/");
+                    buffer.append ("/");
                 else
-                    sb.append (drive).append (":/");
+                    buffer.append (drive).append (":/");
                 break;
 
             // Path segment
             case 2:
-                final String pathSegment = StreamUtils.readWithLengthUTF16 (in);
-                sb.append (pathSegment).append ('/');
+                buffer.append (StreamUtils.readWithLengthUTF16 (in)).append ('/');
                 break;
 
             // Parent directory
             case 3:
-                sb.append ("../");
+                buffer.append ("../");
                 break;
 
             // File name
             case 4:
-                final String filenameSegment = StreamUtils.readWithLengthUTF16 (in);
-                sb.append (filenameSegment);
+                buffer.append (StreamUtils.readWithLengthUTF16 (in));
+                break;
+
+            case 6:
+                // Some kind of top element, maybe a '.'?!
+                break;
+
+            // NKX Filename
+            case 8:
+                buffer.append (StreamUtils.readWithLengthUTF16 (in));
+                break;
+
+            // NKM Filename
+            case 9:
+                buffer.append (StreamUtils.readWithLengthUTF16 (in));
                 break;
 
             default:
                 throw new IOException (Functions.getMessage ("IDS_NKI5_UNSUPPORTED_FILELIST_SEGMENT", Integer.toString (segmentType)));
         }
+    }
+
+
+    /**
+     * Get the special file paths.
+     *
+     * @return The file paths
+     */
+    public List<String> getSpecialFiles ()
+    {
+        return this.specialFiles;
     }
 
 
@@ -268,83 +214,17 @@ public class FileList
      */
     public List<String> getSampleFiles ()
     {
-        return this.sampleFilePaths;
+        return this.sampleFiles;
     }
 
 
     /**
-     * Get the absolute path of the samples.
+     * Get the other file paths.
      *
-     * @return The absolute sample path
+     * @return The file paths
      */
-    public String getAbsoluteSourcePath ()
+    public List<String> getOtherFiles ()
     {
-        return this.absoluteSourcePath;
-    }
-
-
-    /**
-     * Get the absolute monolith source path
-     *
-     * @return The absolute monolith source path
-     */
-    public String getAbsoluteMonolithSourcePath ()
-    {
-        return this.absoluteMonolithSourcePath;
-    }
-
-
-    /**
-     * Get the resource file.
-     *
-     * @return The resource file
-     */
-    public String getResourceFile ()
-    {
-        return this.resourceFile;
-    }
-
-
-    /**
-     * Get the NKI filename.
-     *
-     * @return The NKI filename
-     */
-    public String getNkiFilename ()
-    {
-        return this.nkiFilename;
-    }
-
-
-    /**
-     * Get the wallpaper file name.
-     *
-     * @return The wallpaper file name
-     */
-    public String getWallpaperFilename ()
-    {
-        return this.wallpaperFilename;
-    }
-
-
-    /**
-     * Get the file name of the impulse response.
-     *
-     * @return The file name of the impulse response
-     */
-    public String getImpulseResponseSampleFilename ()
-    {
-        return this.impulseResponseSampleFilename;
-    }
-
-
-    /**
-     * Get the folder which contain the resources.
-     *
-     * @return The resource folder
-     */
-    public String getResourceFolder ()
-    {
-        return this.resourceFolder;
+        return this.otherFiles;
     }
 }